Use simple ASCII double quotes and put commas outside of quotes.
The reasoning is that either it doesn't matter, but if it does, the language model will have seen many more normal double quotes during training.
Also, while putting the comma inside the quote is correct from the English language perspective, it might be clearer for the LLM if we just put in the things we want it to care about.